Apply adversarial loss without sampling
This is tricky: we want to learn a particular latent space, one that
is disentangled in some ideal way so that each dimension is meaningful.

By necessity, this means that the distribution of each dimension E(x)_i
is going to be something potentially very different from Gaussian.
A dimension might be uniform, or one-sided, or binary.

More importantly, the joint distribution P[E(x)] will not be anything
close to a Gaussian. Paddle position and ball position are correlated,
spaceship location and alien location are correlated, etc etc.

So basically the idea here is, if we want to recover the True Latent
Space via disentanglement then we cannot expect that latent space to
be Gaussian; it might be something very complex.

So, we can't sample from a latent space prior. That rules out a standard
GAN or a Makhzani autoencoder, but we can still do an SRGAN-style net.