Replace things with stuff
Here I use a latent space equal in size to the input image.

Previously, I've tried to learn a representation of the world as a set
of things, each thing independent of its position in the grid.

But that's not how convolutional networks operate, and it's not how
neural networks learn: things are inherently symbolic and sparse.
Backpropagation can't learn from a small number of things.

Stuff, on the other hand, is eminently well-suited to representation by
convolutional deep networks. Stuff *is* the grid of places in which it
is embedded: each pixel of stuff interacts locally with its neighbors.

Each pixel is a state machine or a cellular automaton. At each time step
it looks at its neighbors, thinks a little, and then updates its state.