Predict multiple attributed rewards
Each positive or negative reward must is *caused by* a factor.

With a training process similar to weakly-supervised multi-instance
learning, we explicitly encode the reward as being caused by some factor
or combination of factors in the environment.

Additionally, in this environment there are multiple kinds of rewards.
We predict each one independently.