Add counterfactual L1 loss
A good world model should have the property that, given two different
courses of action to simulate, only a minimal number of factors should
be different between the two.

Like, if you went back in time and killed Hitler, then the Beatles
should still have been popular. Just, the lyrics to "Hold your Hand"
would be different, and Paul would have a moustache. All the same
buildings would be built, just they'd be named after different people.

This actually might not be true at all for the real world, but it is a
desirable property to have for a neural network simulation that we're
going to rely on for decision-making.