Add self-control bias term
This prevents learned helplessness in reinforcement learning agents.
Literally, it forces a causal model where actions influence rewards.
With this bias, it is impossible to believe that rewards are outside of
the agent's control. And it's easier than cognitive behavioral therapy.

Note: Langer et al. suggests that the opposite of learned helplessness
is a form of mania. May want to keep in mind when building future AI.